% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jump_cutoff}
\alias{jump_cutoff}
\title{Estimate threshold of maximum p-values across two studies to control FDR.}
\usage{
jump_cutoff(pa_in, pb_in, xi_in, alpha_in)
}
\arguments{
\item{pa_in}{A numeric vector of p-values from study 1.}

\item{pb_in}{A numeric vector of p-values from study 2.}

\item{xi_in}{The estimates of proportions of three null components.}

\item{alpha_in}{The FDR level to control, default is 0.05.}
}
\value{
A list including the maximum of p-values and estimated threshold for FDR control.
}
\description{
Estimate threshold of maximum p-values across two studies to control FDR.
}
