% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JUMP.R
\name{xi.est}
\alias{xi.est}
\title{Estimate the proportions of the composite null hypotheses from the data}
\usage{
xi.est(pvals1, pvals2, lambda = seq(0.01, 0.8, 0.01))
}
\arguments{
\item{pvals1}{a numeric vector of the p-values from study 1}

\item{pvals2}{a numeric vector of the p-values from study 2}

\item{lambda}{the value of the the tuning parameter to estimate}
}
\value{
a list of estimates of the proportions of four groups corresponding to the joint hidden states
}
\description{
The function provides conservative estimates of the proportions defined in the four-group model
}
